+++ src/util/ecdsa.rs
@@ -12,9 +12,8 @@ use secp256k1;
 
 use crate::prelude::*;
 use crate::hashes::hex::{self, FromHex};
-use crate::blockdata::transaction::NonStandardSighashType;
-use crate::EcdsaSighashType;
 use crate::internal_macros::write_err;
+use crate::util::sighash::{EcdsaSighashType, NonStandardSighashType};
 
 /// An ECDSA signature with the corresponding hash type.
 #[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
